Rename selinux parameter
`selinux` is a core fact that returns a boolean.
https://docs.puppet.com/facter/latest/core_facts.html#selinux

If also set as a parameter in Foreman, it becomes the output of the ENC
and overrides the true/false returned by facter breaking puppet modules
including theforeman/foreman that uses a `str2bool($::selinux)`
expression.

https://github.com/theforeman/puppet-foreman/blob/5.2.0/manifests/install.pp#L23

`selinux-mode` is hopefully a good choice for the rename.  It doesn't
conflict with any other core fact such as
https://docs.puppet.com/facter/latest/core_facts.html#selinuxconfigmode
or
https://docs.puppet.com/facter/latest/core_facts.html#selinuxcurrentmode